starting to componentize the header, tools, and terrain
