+++ examples/react/terrain-create-react-app/README.md
@@ -2,19 +2,26 @@
 
 ## About
 
-so what gives?! write stuff here when things have settled...
+After discovering this awesome repo from the blog post, I wanted to play around with this codebase and begin to port it to React and es6. I like the concept of generating a random map but also want to explore tweaking it to my liking so that I might be able to use it in a game or story. Once I have editted the city names and have a map that I like, I should be able to save and export it. 
 
-## STEPS
+In this demo, I'm leveraging Facebook's [create react app](https://facebook.github.io/react/blog/2016/07/22/create-apps-with-no-configuration.html) tool to quickly get a React app up and running that uses the lastest es6 features. Then, I've added the latest version of D3. 
+
+
+## Dev Steps
 
 1. 
 
-[install create react app](https://facebook.github.io/react/blog/2016/07/22/create-apps-with-no-configuration.html)
+npm install
 
 2. 
 
-npm install d3 --save
-
-3. 
-
 npm start 
+
+## Production Step
+
+1. 
+
+npm run build
+
+
