+++ examples/react/terrain-create-react-app/src/boilerplate/Animal.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+
+export default class Animal extends Component {
+
+	constructor(){
+		super();
+	}
+
+	render() {
+		return(
+			<p>hello world</p>
+		);
+	}
+
+}
